// frontend/src/autogen/proto.d.ts
export class UploadedFileInfo implements IUploadedFileInfo {
  /**
   * Constructs a new UploadedFileInfo.
   * @param [properties] Properties to set
   */
  constructor(properties?: IUploadedFileInfo)

  /** UploadedFileInfo id. */
  public id: number | Long

  /** UploadedFileInfo name. */
  public name: string

  /** UploadedFileInfo size. */
  public size: number

  /**
   * Creates a new UploadedFileInfo instance using the specified properties.
   * @param [properties] Properties to set
   * @returns UploadedFileInfo instance
   */
  public static create(properties?: IUploadedFileInfo): UploadedFileInfo

  /**
   * Encodes the specified UploadedFileInfo message. Does not implicitly {@link UploadedFileInfo.verify|verify} messages.
   * @param message UploadedFileInfo message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IUploadedFileInfo,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified UploadedFileInfo message, length delimited. Does not implicitly {@link UploadedFileInfo.verify|verify} messages.
   * @param message UploadedFileInfo message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IUploadedFileInfo,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes an UploadedFileInfo message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns UploadedFileInfo
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): UploadedFileInfo

  /**
   * Decodes an UploadedFileInfo message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns UploadedFileInfo
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): UploadedFileInfo

  /**
   * Verifies an UploadedFileInfo message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates an UploadedFileInfo message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns UploadedFileInfo
   */
  public static fromObject(object: { [k: string]: any }): UploadedFileInfo

  /**
   * Creates a plain object from an UploadedFileInfo message. Also converts values to other types if specified.
   * @param message UploadedFileInfo
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: UploadedFileInfo,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this UploadedFileInfo to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for UploadedFileInfo
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}

/** Properties of a FileUploaderState. */
export interface IFileUploaderState {
  /** FileUploaderState maxFileId */
  maxFileId?: number | Long | null

  /** FileUploaderState uploadedFileInfo */
  uploadedFileInfo?: IUploadedFileInfo[] | null
}

/** Represents a FileUploaderState. */
export class FileUploaderState implements IFileUploaderState {
  /**
   * Constructs a new FileUploaderState.
   * @param [properties] Properties to set
   */
  constructor(properties?: IFileUploaderState)

  /** FileUploaderState maxFileId. */
  public maxFileId: number | Long

  /** FileUploaderState uploadedFileInfo. */
  public uploadedFileInfo: IUploadedFileInfo[]

  /**
   * Creates a new FileUploaderState instance using the specified properties.
   * @param [properties] Properties to set
   * @returns FileUploaderState instance
   */
  public static create(properties?: IFileUploaderState): FileUploaderState

  /**
   * Encodes the specified FileUploaderState message. Does not implicitly {@link FileUploaderState.verify|verify} messages.
   * @param message FileUploaderState message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IFileUploaderState,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified FileUploaderState message, length delimited. Does not implicitly {@link FileUploaderState.verify|verify} messages.
   * @param message FileUploaderState message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IFileUploaderState,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes a FileUploaderState message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns FileUploaderState
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): FileUploaderState

  /**
   * Decodes a FileUploaderState message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns FileUploaderState
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): FileUploaderState

  /**
   * Verifies a FileUploaderState message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates a FileUploaderState message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns FileUploaderState
   */
  public static fromObject(object: { [k: string]: any }): FileUploaderState

  /**
   * Creates a plain object from a FileUploaderState message. Also converts values to other types if specified.
   * @param message FileUploaderState
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: FileUploaderState,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this FileUploaderState to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for FileUploaderState
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}
// ...
export interface IFileUploader {
  /** FileUploader id */
  id?: string | null

  /** FileUploader label */
  label?: string | null

  /** FileUploader type */
  type?: string[] | null

  /** FileUploader maxUploadSizeMb */
  maxUploadSizeMb?: number | null

  /** FileUploader multipleFiles */
  multipleFiles?: boolean | null

  /** FileUploader help */
  help?: string | null

  /** FileUploader formId */
  formId?: string | null

  /** FileUploader disabled */
  disabled?: boolean | null

  /** FileUploader labelVisibility */
  labelVisibility?: ILabelVisibilityMessage | null
}

/** Represents a FileUploader. */
export class FileUploader implements IFileUploader {
  /**
   * Constructs a new FileUploader.
   * @param [properties] Properties to set
   */
  constructor(properties?: IFileUploader)

  /** FileUploader id. */
  public id: string

  /** FileUploader label. */
  public label: string

  /** FileUploader type. */
  public type: string[]

  /** FileUploader maxUploadSizeMb. */
  public maxUploadSizeMb: number

  /** FileUploader multipleFiles. */
  public multipleFiles: boolean

  /** FileUploader help. */
  public help: string

  /** FileUploader formId. */
  public formId: string

  /** FileUploader disabled. */
  public disabled: boolean

  /** FileUploader labelVisibility. */
  public labelVisibility?: ILabelVisibilityMessage | null

  /**
   * Creates a new FileUploader instance using the specified properties.
   * @param [properties] Properties to set
   * @returns FileUploader instance
   */
  public static create(properties?: IFileUploader): FileUploader

  /**
   * Encodes the specified FileUploader message. Does not implicitly {@link FileUploader.verify|verify} messages.
   * @param message FileUploader message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encode(
    message: IFileUploader,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Encodes the specified FileUploader message, length delimited. Does not implicitly {@link FileUploader.verify|verify} messages.
   * @param message FileUploader message or plain object to encode
   * @param [writer] Writer to encode to
   * @returns Writer
   */
  public static encodeDelimited(
    message: IFileUploader,
    writer?: $protobuf.Writer
  ): $protobuf.Writer

  /**
   * Decodes a FileUploader message from the specified reader or buffer.
   * @param reader Reader or buffer to decode from
   * @param [length] Message length if known beforehand
   * @returns FileUploader
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decode(
    reader: $protobuf.Reader | Uint8Array,
    length?: number
  ): FileUploader

  /**
   * Decodes a FileUploader message from the specified reader or buffer, length delimited.
   * @param reader Reader or buffer to decode from
   * @returns FileUploader
   * @throws {Error} If the payload is not a reader or valid buffer
   * @throws {$protobuf.util.ProtocolError} If required fields are missing
   */
  public static decodeDelimited(
    reader: $protobuf.Reader | Uint8Array
  ): FileUploader

  /**
   * Verifies a FileUploader message.
   * @param message Plain object to verify
   * @returns `null` if valid, otherwise the reason why it is not
   */
  public static verify(message: { [k: string]: any }): string | null

  /**
   * Creates a FileUploader message from a plain object. Also converts values to their respective internal types.
   * @param object Plain object
   * @returns FileUploader
   */
  public static fromObject(object: { [k: string]: any }): FileUploader

  /**
   * Creates a plain object from a FileUploader message. Also converts values to other types if specified.
   * @param message FileUploader
   * @param [options] Conversion options
   * @returns Plain object
   */
  public static toObject(
    message: FileUploader,
    options?: $protobuf.IConversionOptions
  ): { [k: string]: any }

  /**
   * Converts this FileUploader to JSON.
   * @returns JSON object
   */
  public toJSON(): { [k: string]: any }

  /**
   * Gets the default type url for FileUploader
   * @param [typeUrlPrefix] your custom typeUrlPrefix(default "type.googleapis.com")
   * @returns The default type url
   */
  public static getTypeUrl(typeUrlPrefix?: string): string
}
// ...
export interface ILabelVisibilityMessage {
  /** LabelVisibilityMessage value */
  value?: LabelVisibilityMessage.LabelVisibilityOptions | null
}
// ...
export namespace LabelVisibilityMessage {
  /** LabelVisibilityOptions enum. */
  enum LabelVisibilityOptions {
    VISIBLE = 0,
    HIDDEN = 1,
    COLLAPSED = 2,
  }
}